Add atomic_make_stt

+++ test/test_decoders/420a/420a.cpp
@@ -7,7 +7,7 @@
 static void test_getCrankAngle(void)
 {
   extern decoder_status_t decoderStatus;
-  extern volatile unsigned long toothLastToothTime;
+  extern volatile uint32_t toothLastToothTime;
   extern volatile int toothCurrentCount;
 
   auto decoder = triggerSetup_420a();
